+++ packages/navigation/docs/components/AkContainerTitleDropdown.js
@@ -2,7 +2,8 @@ import React from 'react';
 
 export const description = (
   <div>
-    This component is designed to render AkContainerTitle and a ExpandIcon.
+    This component is designed to render an AkContainerTitle as a dropdown trigger.
+    Children need to be DropdownItems.
   </div>
 );