+++ packages/navigation/src/components/styled/ContainerTitleIcon.jsx
@@ -5,20 +5,20 @@ const size = gridSize * 5;
 const borderRadius = 4;
 
 const ContainerTitleIcon = styled.div`
-  width: ${size}px;
+  align-items: center;
+  display: flex;
   height: ${size}px;
-  
-  /* fix a flexbox issue in IE11 */
-  min-width: ${size}px;
+  justify-content: center;
+  width: ${size}px;
 
   /* We need to ensure that any image passed in as a child (<img/>, <svg/>
   etc.) receives the correct width, height and border radius. We don't
   currently assume that the image passed in is the correct dimensions, or has
   width / height 100% */
-  & > * {
-    width: ${size}px;
-    height: ${size}px;
+  & > img {
     border-radius: ${borderRadius}px;
+    height: ${size}px;
+    width: ${size}px;
   }
 `;