+++ packages/single-select/stories/00-select-overview-story.jsx
@@ -29,6 +29,8 @@ import SelectWithHeapsOfOptions from './examples/SelectWithHeapsOfOptions';
 import SelectWithHeapsOfOptionsRaw from '!raw!./examples/SelectWithHeapsOfOptions';
 import SelectWithDescriptions from './examples/SelectWithDescriptions';
 import SelectWithDescriptionsRaw from '!raw!./examples/SelectWithDescriptions';
+import SelectWithFilterValues from './examples/SelectWithFilterValues';
+import SelectWithFilterValuesRaw from '!raw!./examples/SelectWithFilterValues';
 import SelectWithLongDescriptions from './examples/SelectWithLongDescriptions';
 import SelectWithLongDescriptionsRaw from '!raw!./examples/SelectWithLongDescriptions';
 import SelectWithTooltips from './examples/SelectWithTooltips';
@@ -39,6 +41,12 @@ import SelectOpenedByDefault from './examples/SelectOpenedByDefault';
 import SelectOpenedByDefaultRaw from '!raw!./examples/SelectOpenedByDefault';
 import SelectNoPositionFlip from './examples/SelectNoPositionFlip';
 import SelectNoPositionFlipRaw from '!raw!./examples/SelectNoPositionFlip';
+
+import ShouldFocus from './examples/ShouldFocus';
+import ShouldFocusRaw from '!raw!./examples/ShouldFocus';
+
+import ShouldFocusStateful from './examples/ShouldFocusStateful';
+import ShouldFocusStatefulRaw from '!raw!./examples/ShouldFocusStateful';
 /* eslint-enable import/first, import/no-duplicates */
 
 // Dummy components exist so that we have a component to pass to <Props/>
@@ -163,6 +171,28 @@ storiesOf(name, module)
       <Props component={Select} descriptions={propDescriptions} types={propTypes} />
     </Chrome>
   ))
+  .add('Single select with should focus', () => (
+    <Chrome title="With should focus">
+      <Description>
+        <p style={{ marginBottom: '12px' }}>Should focus</p>
+      </Description>
+      {ShouldFocus}
+      <Code>
+        {ShouldFocusRaw}
+      </Code>
+    </Chrome>
+  ))
+  .add('Single select (stateful) with should focus', () => (
+    <Chrome title="With should focus">
+      <Description>
+        <p style={{ marginBottom: '12px' }}>Should focus (stateful)</p>
+      </Description>
+      {ShouldFocusStateful}
+      <Code>
+        {ShouldFocusStatefulRaw}
+      </Code>
+    </Chrome>
+  ))
   .add('Select that fills all available space', () => (
     <Chrome title="Select that fills all available space">
       {WideSelect}
@@ -254,6 +284,17 @@ storiesOf(name, module)
       {SelectWithAutocomplete}
       <Code>
         {SelectWithAutocompleteRaw}
+      </Code>
+      <Props component={Select} descriptions={propDescriptions} types={propTypes} />
+    </Chrome>
+  ))
+  .add('Select with filter values', () => (
+    <Chrome title="Select with filter values">
+      <div style={{ width: '300px' }}>
+        {SelectWithFilterValues}
+      </div>
+      <Code>
+        {SelectWithFilterValuesRaw}
       </Code>
       <Props component={Select} descriptions={propDescriptions} types={propTypes} />
     </Chrome>