+++ packages/breadcrumbs/docs/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/breadcrumbs
 
+## 2.5.1 (2017-08-21)
+
+* bug fix; fix PropTypes warning ([040d579](https://bitbucket.org/atlassian/atlaskit/commits/040d579))
 ## 2.5.0 (2017-08-17)